abstract nominator status items
