+++ src/pages/Nominate/Active/Status/index.tsx
@@ -1,201 +1,18 @@
 // Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import {
-  faBolt,
-  faChevronCircleRight,
-  faGear,
-  faSignOutAlt,
-  faWallet,
-} from '@fortawesome/free-solid-svg-icons';
-import { useApi } from 'contexts/Api';
-import { useBalances } from 'contexts/Balances';
-import { useConnect } from 'contexts/Connect';
-import { useFastUnstake } from 'contexts/FastUnstake';
-import { useModal } from 'contexts/Modal';
-import { useNetworkMetrics } from 'contexts/Network';
-import { useSetup } from 'contexts/Setup';
-import { useStaking } from 'contexts/Staking';
-import { useUi } from 'contexts/UI';
 import { CardWrapper } from 'library/Graphs/Wrappers';
-import { useNominationStatus } from 'library/Hooks/useNominationStatus';
-import { PayeeItem, usePayeeConfig } from 'library/Hooks/usePayeeConfig';
-import { useUnstaking } from 'library/Hooks/useUnstaking';
-import { Stat } from 'library/Stat';
-import { useTranslation } from 'react-i18next';
-import { clipAddress } from 'Utils';
 import { Separator } from 'Wrappers';
+import { ControllerStatus } from './ControllerStatus';
+import { NominationStatus } from './NominationStatus';
+import { PayoutDestinationStatus } from './PayoutDestinationStatus';
 
-export const Status = ({ height }: { height: number }) => {
-  const { t } = useTranslation();
-  const { isSyncing } = useUi();
-  const { openModalWith } = useModal();
-  const { isReady } = useApi();
-  const { getBondedAccount } = useBalances();
-  const { metrics } = useNetworkMetrics();
-  const { activeAccount, isReadOnlyAccount, getAccount } = useConnect();
-  const { setOnNominatorSetup, getNominatorSetupPercent }: any = useSetup();
-  const { getNominationsStatus, staking, inSetup, hasController } =
-    useStaking();
-  const { checking, isExposed } = useFastUnstake();
-  const { getFastUnstakeText, isUnstaking, isFastUnstaking } = useUnstaking();
-  const controller = getBondedAccount(activeAccount);
-  const { getNominationStatus } = useNominationStatus();
-  const { getPayeeItems } = usePayeeConfig();
-  const { fastUnstakeErasToCheckPerBlock } = metrics;
-  const { payee } = staking;
-
-  // get nomination status
-  const nominationStatuses = getNominationsStatus();
-
-  // get active nominations
-  const activeNominees = Object.entries(nominationStatuses)
-    .map(([k, v]: any) => (v === 'active' ? k : false))
-    .filter((v) => v !== false);
-
-  const getPayeeStatus = () => {
-    if (inSetup()) {
-      return t('nominate.notAssigned', { ns: 'pages' });
-    }
-    const status = getPayeeItems(true).find(
-      (p: PayeeItem) => p.value === payee.destination
-    )?.activeTitle;
-
-    if (status) {
-      return status;
-    }
-    return t('nominate.notAssigned', { ns: 'pages' });
-  };
-
-  const payeeIcon = inSetup()
-    ? undefined
-    : getPayeeItems(true).find((p: PayeeItem) => p.value === payee.destination)
-        ?.icon || faWallet;
-
-  let startTitle = t('nominate.startNominating', { ns: 'pages' });
-  if (inSetup()) {
-    const progress = getNominatorSetupPercent(activeAccount);
-    if (progress > 0) {
-      startTitle += `: ${progress}%`;
-    }
-  }
-
-  const fastUnstakeText = getFastUnstakeText();
-  const regularUnstakeButton = {
-    title: t('nominate.unstake', { ns: 'pages' }),
-    icon: faSignOutAlt,
-    disabled: !isReady || isReadOnlyAccount(controller) || !activeAccount,
-    onClick: () => openModalWith('Unstake', {}, 'small'),
-  };
-
-  const fastUnstakeButton = {
-    disabled: checking,
-    title: fastUnstakeText,
-    icon: faBolt,
-    onClick: () => {
-      openModalWith('ManageFastUnstake', {}, 'small');
-    },
-  };
-
-  const unstakeButton =
-    fastUnstakeErasToCheckPerBlock > 0 &&
-    !activeNominees.length &&
-    (checking || !isExposed)
-      ? fastUnstakeButton
-      : regularUnstakeButton;
-
-  return (
-    <CardWrapper height={height}>
-      <Stat
-        label={t('nominate.status', { ns: 'pages' })}
-        helpKey="Nomination Status"
-        stat={getNominationStatus(activeAccount, 'nominator').message}
-        buttons={
-          !inSetup()
-            ? !isUnstaking && !isReadOnlyAccount(controller)
-              ? [unstakeButton]
-              : []
-            : [
-                {
-                  title: startTitle,
-                  icon: faChevronCircleRight,
-                  transform: 'grow-1',
-                  large: true,
-                  disabled:
-                    !isReady ||
-                    isReadOnlyAccount(activeAccount) ||
-                    !activeAccount,
-                  onClick: () => setOnNominatorSetup(true),
-                },
-              ]
-        }
-      />
-      <Separator />
-      <Stat
-        label="Payout Destination"
-        helpKey="Payout Destination"
-        icon={payeeIcon}
-        stat={getPayeeStatus()}
-        buttons={
-          !inSetup()
-            ? [
-                {
-                  title: t('nominate.update', { ns: 'pages' }),
-                  icon: faGear,
-                  small: true,
-                  disabled:
-                    inSetup() ||
-                    isSyncing ||
-                    isReadOnlyAccount(activeAccount) ||
-                    isFastUnstaking,
-                  onClick: () => openModalWith('UpdatePayee', {}, 'small'),
-                },
-              ]
-            : []
-        }
-      />
-      <Separator />
-      <Stat
-        label={t('nominate.controllerAccount', { ns: 'pages' })}
-        helpKey="Stash and Controller Accounts"
-        stat={
-          controller
-            ? {
-                address: controller,
-                display:
-                  getAccount(controller)?.name || clipAddress(controller),
-              }
-            : `${t('nominate.none', { ns: 'pages' })}`
-        }
-        buttons={
-          !inSetup()
-            ? [
-                {
-                  title: t('nominate.change', { ns: 'pages' }),
-                  icon: faGear,
-                  small: true,
-                  disabled:
-                    !isReady ||
-                    !hasController() ||
-                    isReadOnlyAccount(activeAccount) ||
-                    isFastUnstaking,
-                  onClick: () => openModalWith('UpdateController', {}, 'large'),
-                },
-              ]
-            : []
-        }
-        copy={
-          !controller
-            ? undefined
-            : {
-                content: controller,
-                notification: {
-                  title: t('nominate.addressCopied', { ns: 'pages' }),
-                  subtitle: controller,
-                },
-              }
-        }
-      />
-    </CardWrapper>
-  );
-};
+export const Status = ({ height }: { height: number }) => (
+  <CardWrapper height={height}>
+    <NominationStatus />
+    <Separator />
+    <PayoutDestinationStatus />
+    <Separator />
+    <ControllerStatus />
+  </CardWrapper>
+);
